feat: move tasks to dedicated folder
